+++ src/controllers/component.controller.js
@@ -2,16 +2,11 @@
 require('dotenv').config();
 //Services
 const {
-  // updateComponentService,
-  deleteComponentService,
-  getAllComponentService,
   getComponentByIdService,
   getAllComponentLikeCodeService,
-  getAllWithAttributesComponentService,
   getAllComponentLikeImageService,
   getAllComponentLikePartNumberService,
   getAllComponentLikeCategoryAndMakerService,
-  getAllComponentWithDetailsService,
   getAllComponentWithBipolarTransistorService,
   getAllComponentWithAllModelsService,
   getAllComponentLikeDescriptionService,
@@ -48,346 +43,8 @@ const ORDER_BY_DESCRIPTION_VALUE_ERROR =
   paginationDescriptionValueError.ORDER_BY_DESCRIPTION_VALUE_ERROR;
 const ORDER_AT_DESCRIPTION_VALUE_ERROR =
   paginationDescriptionValueError.ORDER_AT_DESCRIPTION_VALUE_ERROR;
-let deletedComponent;
 let msgResponse;
 let msgLog;
-
-
-/**
- * @description delete a component from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const deleteComponentController = async (req, res) => {
-  try {
-    msgResponse = null;
-    msgLog = null;
-    deletedComponent = await deleteComponentService(req);
-
-    switch (deletedComponent) {
-      case CONNECTION_ERROR_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_ERROR_STATUS_DETAIL });
-        break;
-      case CONNECTION_REFUSED_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_REFUSED_STATUS_DETAIL });
-        break;
-      case 0:
-      case undefined:
-      case null:
-        res
-          .status(BAD_REQUEST_CODE)
-          .send({ error: 'Bad request, could not delete a component.' });
-        break;
-      default:
-        if (
-          typeof deletedComponent === 'object' &&
-          deletedComponent.hasOwnProperty('objectDeleted')
-        ) {
-          res.status(OK_CODE).send(deletedComponent);
-          break;
-        }
-        res.status(BAD_REQUEST_CODE).send({ error: deletedComponent });
-        break;
-    }
-  } catch (error) {
-    msgResponse = 'ERROR in deleteComponentController() function.';
-    msgLog = msgResponse + `Caused by ${error}`;
-    console.log(msgLog);
-    res.status(INTERNAL_SERVER_ERROR_CODE).send({ error: msgResponse });
-  }
-};
-
-/**
- * @description get all paginated components to database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllComponentController = async (req, res) => {
-  try {
-    msgResponse = null;
-    msgLog = null;
-
-    componentList = await getAllComponentService(req);
-
-    switch (componentList) {
-      case CONNECTION_ERROR_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_ERROR_STATUS_DETAIL });
-        break;
-      case CONNECTION_REFUSED_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_REFUSED_STATUS_DETAIL });
-        break;
-      case ORDER_BY_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_BY_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case ORDER_AT_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_AT_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case undefined:
-      case null:
-        res.status(BAD_REQUEST_CODE).send({
-          error: 'Bad request, failed to obtain paginated component list.',
-        });
-        break;
-      default:
-        if (
-          (typeof componentList === 'object' &&
-            componentList[0]?.hasOwnProperty('id')) ||
-          (Array.isArray(componentList) && componentList.length)
-        ) {
-          res.status(OK_CODE).send(componentList);
-          break;
-        } else if (
-          (typeof componentList === 'object' &&
-            Object.keys(componentList).length == 0) ||
-          (Array.isArray(componentList) && componentList.length == 0)
-        ) {
-          res
-            .status(OK_CODE)
-            .send({ ok: 'No items found according to all atributes.' });
-        } else {
-          res.status(BAD_REQUEST_CODE).send({ error: componentList });
-          break;
-        }
-    }
-  } catch (error) {
-    msgResponse = `Error in getAllComponentController() function. Caused by ${error}`;
-    msgLog = msgResponse + `Caused by ${error}`;
-    console.log(msgLog);
-    res.status(INTERNAL_SERVER_ERROR_CODE).send({ error: msgResponse });
-  }
-};
-
-/**
- * @description get all paginated components list according to all attributes from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllWithAttributesComponentController = async (req, res) => {
-  try {
-    msgResponse = null;
-    msgLog = null;
-    componentList = null;
-
-    componentList = await getAllWithAttributesComponentService(req);
-
-    switch (componentList) {
-      case CONNECTION_ERROR_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_ERROR_STATUS_DETAIL });
-        break;
-      case CONNECTION_REFUSED_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_REFUSED_STATUS_DETAIL });
-        break;
-      case ORDER_BY_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_BY_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case ORDER_AT_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_AT_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case undefined:
-      case null:
-        res.status(BAD_REQUEST_CODE).send({
-          error:
-            'Bad request, failed to get all paginated components list according to all attributes.',
-        });
-        break;
-      default:
-        if (
-          (typeof componentList === 'object' &&
-            componentList[0]?.hasOwnProperty('id')) ||
-          (Array.isArray(componentList) && componentList.length)
-        ) {
-          res.status(OK_CODE).send(componentList);
-          break;
-        } else if (
-          (typeof componentList === 'object' &&
-            Object.keys(componentList).length == 0) ||
-          (Array.isArray(componentList) && componentList.length == 0)
-        ) {
-          res
-            .status(OK_CODE)
-            .send({ ok: 'No items found according to all atributes.' });
-        } else {
-          res.status(BAD_REQUEST_CODE).send({ error: componentList });
-          break;
-        }
-    }
-  } catch (error) {
-    msgResponse =
-      'ERROR in getAllWithAttributesComponentController() function.';
-    msgLog = msgResponse + `Caused by ${error}`;
-    console.log(msgLog);
-    res.status(INTERNAL_SERVER_ERROR_CODE).send({ error: msgResponse });
-  }
-};
-
-/**
- * @description get all paginated components list with details for component_details table from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllWithDetailComponentController = async (req, res) => {
-  try {
-    msgResponse = null;
-    msgLog = null;
-
-    componentList = await getAllComponentWithDetailsService(req);
-
-    switch (componentList) {
-      case CONNECTION_ERROR_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_ERROR_STATUS_DETAIL });
-        break;
-      case CONNECTION_REFUSED_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_REFUSED_STATUS_DETAIL });
-        break;
-      case ORDER_BY_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_BY_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case ORDER_AT_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_AT_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case undefined:
-      case null:
-        res.status(BAD_REQUEST_CODE).send({
-          error:
-            'Bad request, failed to get all paginated components list and components_details according to all attributes.',
-        });
-        break;
-      default:
-        if (
-          (typeof componentList === 'object' &&
-            componentList[0]?.hasOwnProperty('id')) ||
-          (Array.isArray(componentList) && componentList.length)
-        ) {
-          res.status(OK_CODE).send(componentList);
-          break;
-        } else if (
-          (typeof componentList === 'object' &&
-            Object.keys(componentList).length == 0) ||
-          (Array.isArray(componentList) && componentList.length == 0)
-        ) {
-          res.status(OK_CODE).send({
-            ok: 'No items found according to the component details model.',
-          });
-        } else {
-          res.status(BAD_REQUEST_CODE).send({ error: componentList });
-          break;
-        }
-    }
-  } catch (error) {
-    msgResponse = 'ERROR in getAllWithDetailComponentController() function.';
-    msgLog = msgResponse + `Caused by ${error}`;
-    console.log(msgLog);
-    res.status(INTERNAL_SERVER_ERROR_CODE).send({ error: msgResponse });
-  }
-};
-
-/**
- * @description get all paginated components list with details for bipolar_transistor table from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllWithBipolarTransistorComponentController = async (req, res) => {
-  try {
-    msgResponse = null;
-    msgLog = null;
-
-    componentList = await getAllComponentWithBipolarTransistorService(req);
-
-    switch (componentList) {
-      case CONNECTION_ERROR_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_ERROR_STATUS_DETAIL });
-        break;
-      case CONNECTION_REFUSED_STATUS:
-        res
-          .status(INTERNAL_SERVER_ERROR_CODE)
-          .send({ error: CONNECTION_REFUSED_STATUS_DETAIL });
-        break;
-      case ORDER_BY_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_BY_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case ORDER_AT_NAME_VALUE_ERROR:
-        res.status(BAD_REQUEST_CODE).send({
-          error: ORDER_AT_DESCRIPTION_VALUE_ERROR,
-        });
-        break;
-      case undefined:
-      case null:
-        res.status(BAD_REQUEST_CODE).send({
-          error:
-            'Bad request, failed to get all paginated components list and bipolar_transistor according to all attributes.',
-        });
-        break;
-      default:
-        if (
-          (typeof componentList === 'object' &&
-            componentList[0]?.hasOwnProperty('id')) ||
-          (Array.isArray(componentList) && componentList.length)
-        ) {
-          res.status(OK_CODE).send(componentList);
-          break;
-        } else if (
-          (typeof componentList === 'object' &&
-            Object.keys(componentList).length == 0) ||
-          (Array.isArray(componentList) && componentList.length == 0)
-        ) {
-          res.status(OK_CODE).send({
-            ok: 'No items found according to the bipolar transistor model.',
-          });
-        } else {
-          res.status(BAD_REQUEST_CODE).send({ error: componentList });
-          break;
-        }
-    }
-  } catch (error) {
-    msgResponse =
-      'ERROR in getAllWithBipolarTransistorComponentController() function.';
-    msgLog = msgResponse + `Caused by ${error}`;
-    console.log(msgLog);
-    res.status(INTERNAL_SERVER_ERROR_CODE).send({ error: msgResponse });
-  }
-};
 
 /**
  * @description get all paginated components list with all models/table from the database
@@ -1314,11 +971,6 @@ const getAllComponentLikePriceMinMaxController = async (req, res) => {
 };
 
 module.exports = {
-  deleteComponentController,
-  getAllComponentController,
-  getAllWithAttributesComponentController,
-  getAllWithDetailComponentController,
-  getAllWithBipolarTransistorComponentController,
   getAllWithAllModelsComponentController,
   getComponentByIdController,
   getAllComponentLikeCodeController,
