+++ src/services/component.service.js
@@ -36,7 +36,6 @@ let categoriaParam;
 let descripcionParam;
 let fabricanteParam;
 let stockParam;
-let precioParam;
 let stockMin;
 let stockMax;
 let priceParam;
@@ -49,427 +48,6 @@ let orderBy;
 let orderAt;
 let order;
 let msg;
-
-
-/**
- * @description delete a component from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const deleteComponentService = async (req, res) => {
-  try {
-    //Params
-    deleteComponent = null;
-    msg = null;
-    params = null;
-    idParam = 0;
-
-    //-- start with params ---
-    params = req.params;
-
-    if (params != (null && undefined)) {
-      idParam = params.id ? parseInt(params.id) : null;
-    }
-    //-- end with params  ---
-
-    if (Component != (null && undefined) && idParam != null) {
-      await Component.destroy({
-        where: {
-          id: idParam,
-        },
-      })
-        .then(async (componentItem) => {
-          deleteComponent =
-            componentItem == 1
-              ? {
-                  objectDeleted: `Se ha eliminado correctamente el componente según el id ${idParam}`,
-                }
-              : {
-                  objectDeleted: `No se ha eliminado el componente según el id ${idParam}. Comprobar si el componente existe en la db.`,
-                };
-        })
-        .catch(async (error) => {
-          msg = `Error in deleteComponentService() function when trying to delete a component. Caused by ${error}`;
-          console.log(msg);
-
-          deleteComponent = await checkErrors(error, error.name);
-        });
-    } else {
-      deleteComponent = await checkErrors(null, statusName.CONNECTION_REFUSED);
-    }
-  } catch (error) {
-    msg = `Error in deleteComponentService() function. Caused by ${error}`;
-    console.log(msg);
-    deleteComponent = await checkErrors(error, statusName.CONNECTION_ERROR);
-  }
-  return deleteComponent;
-};
-
-/**
- * @description get all paginated components from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllComponentService = async (req, res) => {
-  try {
-    //Params
-    componentList = null;
-    queryStrParams = null;
-    //Pagination
-    pageSizeNro = 5;
-    pageNro = 0;
-    orderBy = 'id';
-    orderAt = 'ASC';
-    msg = null;
-
-    //-- start with pagination  ---
-    queryStrParams = req.query;
-
-    if (queryStrParams != (null && undefined)) {
-      pageSizeNro = queryStrParams.limit
-        ? parseInt(await queryStrParams.limit)
-        : pageSizeNro;
-      pageNro = queryStrParams.page
-        ? parseInt(await queryStrParams.page)
-        : pageNro;
-      orderBy = queryStrParams.orderBy ? queryStrParams.orderBy : orderBy;
-      orderAt = queryStrParams.orderAt ? queryStrParams.orderAt : orderAt;
-    }
-
-    orderBy = await checkOrderBy(orderBy);
-
-    if (orderBy == (null || undefined)) {
-      return ORDER_BY_NAME_VALUE_ERROR;
-    }
-
-    orderAt = await checkOrderAt(orderAt);
-
-    if (orderAt == (undefined || null)) {
-      return ORDER_AT_NAME_VALUE_ERROR;
-    }
-
-    order = [[orderBy, orderAt]];
-
-    //-- end with pagination  ---
-
-    if (Component != (null && undefined)) {
-      await Component.findAll({
-        attributes: {},
-        limit: pageSizeNro,
-        offset: pageNro,
-        order: order,
-        raw: true,
-      })
-        .then(async (componentItems) => {
-          componentList = componentItems;
-        })
-        .catch(async (error) => {
-          msg = `Error in getAllComponentService() function when trying to get all paginated components. Caused by ${error}`;
-          console.log(msg);
-          componentList = await checkErrors(error, error.name);
-        });
-    } else {
-      componentList = await checkErrors(null, statusName.CONNECTION_REFUSED);
-    }
-  } catch (error) {
-    msg = `Error in getAllComponentService() function. Caused by ${error}`;
-    console.log(msg);
-    componentList = await checkErrors(error, statusName.CONNECTION_ERROR);
-  }
-  return componentList;
-};
-
-/**
- * @description get all paginated components list according to all attributes from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllWithAttributesComponentService = async (req, res) => {
-  try {
-    //Params
-    componentList = null;
-    queryStrParams = null;
-    codigoParam = null;
-    imagenParam = null;
-    nroPiezaParam = null;
-    categoriaParam = null;
-    descripcionParam = null;
-    fabricanteParam = null;
-    stockParam = 0;
-    precioParam = 0;
-    //Pagination
-    pageSizeNro = 5;
-    pageNro = 0;
-    orderBy = 'id';
-    orderAt = 'ASC';
-    msg = null;
-
-    //-- start with querys params and pagination  ---
-    queryStrParams = req.query;
-
-    if (queryStrParams != (null && undefined)) {
-      codigoParam = queryStrParams.codigo ? queryStrParams.codigo : codigoParam;
-      imagenParam = queryStrParams.imagen ? queryStrParams.imagen : imagenParam;
-      nroPiezaParam = queryStrParams.nroPieza
-        ? queryStrParams.nroPieza
-        : nroPiezaParam;
-      categoriaParam = queryStrParams.categoria
-        ? queryStrParams.categoria
-        : categoriaParam;
-      descripcionParam = queryStrParams.descripcion
-        ? queryStrParams.descripcion
-        : descripcionParam;
-      fabricanteParam = queryStrParams.fabricante
-        ? queryStrParams.fabricante
-        : fabricanteParam;
-      stockParam = queryStrParams.stock
-        ? parseInt(queryStrParams.stock)
-        : stockParam;
-      precioParam = queryStrParams.precio ? queryStrParams.precio : precioParam;
-      //Pagination
-      pageSizeNro = queryStrParams.limit
-        ? parseInt(queryStrParams.limit)
-        : pageSizeNro;
-      pageNro = queryStrParams.page ? parseInt(queryStrParams.page) : pageNro;
-      orderBy = queryStrParams.orderBy ? queryStrParams.orderBy : orderBy;
-      orderAt = queryStrParams.orderAt ? queryStrParams.orderAt : orderAt;
-    }
-
-    orderBy = await checkOrderBy(orderBy);
-
-    if (orderBy == (null || undefined)) {
-      return ORDER_BY_NAME_VALUE_ERROR;
-    }
-
-    orderAt = await checkOrderAt(orderAt);
-
-    if (orderAt == (undefined || null)) {
-      return ORDER_AT_NAME_VALUE_ERROR;
-    }
-
-    order = [[orderBy, orderAt]];
-
-    //-- end with querys params and pagination  ---
-
-    if (Component != (null && undefined)) {
-      await Component.findAll({
-        attributes: {},
-        where: {
-          [Op.or]: {
-            codigo: {
-              [Op.like]: `%${codigoParam}%`,
-            },
-            imagen: {
-              [Op.like]: `%${imagenParam}%`,
-            },
-            nro_pieza: {
-              [Op.like]: `%${nroPiezaParam}%`,
-            },
-            categoria: {
-              [Op.like]: `%${categoriaParam}%`,
-            },
-            descripcion: {
-              [Op.like]: `%${descripcionParam}%`,
-            },
-            fabricante: {
-              [Op.like]: `%${fabricanteParam}%`,
-            },
-            stock: {
-              [Op.eq]: stockParam,
-            },
-            precio: {
-              [Op.eq]: precioParam,
-            },
-          },
-        },
-        limit: pageSizeNro,
-        offset: pageNro,
-        order: order,
-        raw: true,
-      })
-        .then(async (componentItems) => {
-          componentList = componentItems;
-        })
-        .catch(async (error) => {
-          msg = `Error in getAllWithAttributesComponentService() function when trying to get all paginated component by all attributes. Caused by ${error}`;
-          console.log(msg);
-          componentList = await checkErrors(error, error.name);
-        });
-    } else {
-      componentList = await checkErrors(null, statusName.CONNECTION_REFUSED);
-    }
-  } catch (error) {
-    msg = `Error in getAllWithAttributesComponentService() function. Caused by ${error}`;
-    console.log(msg);
-    componentList = await checkErrors(error, statusName.CONNECTION_ERROR);
-  }
-  return componentList;
-};
-
-/**
- * @description get all paginated components with component_details from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllComponentWithDetailsService = async (req, res) => {
-  try {
-    //params
-    componentList = null;
-    queryStrParams = null;
-    //Pagination
-    pageSizeNro = 5;
-    pageNro = 0;
-    orderBy = 'id';
-    orderAt = 'ASC';
-    msg = null;
-
-    //-- start with pagination  ---
-    queryStrParams = req.query;
-
-    if (queryStrParams != (null && undefined)) {
-      //Pagination
-      pageSizeNro = queryStrParams.limit
-        ? parseInt(queryStrParams.limit)
-        : pageSizeNro;
-      pageNro = queryStrParams.page ? parseInt(queryStrParams.page) : pageNro;
-      orderBy = queryStrParams.orderBy ? queryStrParams.orderBy : orderBy;
-      orderAt = queryStrParams.orderAt ? queryStrParams.orderAt : orderAt;
-    }
-
-    orderBy = await checkOrderBy(orderBy);
-
-    if (orderBy == (null || undefined)) {
-      return ORDER_BY_NAME_VALUE_ERROR;
-    }
-
-    orderAt = await checkOrderAt(orderAt);
-
-    if (orderAt == (undefined || null)) {
-      return ORDER_AT_NAME_VALUE_ERROR;
-    }
-
-    order = [[orderBy, orderAt]];
-    //-- end with pagination  ---
-
-    if (Component != (null && undefined)) {
-      await Component.findAll({
-        attributes: {},
-        include: [{ model: ComponentDetail, required: true }],
-        limit: pageSizeNro,
-        offset: pageNro,
-        order: order,
-        raw: true,
-        nest: true,
-      })
-        .then(async (componentItems) => {
-          componentList = componentItems;
-        })
-        .catch(async (error) => {
-          msg = `Error in getAllComponentWithDetailsService() function when trying to get all paginated components. Caused by ${error}`;
-          console.log(msg);
-
-          componentList = await checkErrors(error, error.name);
-        });
-    } else {
-      componentList = await checkErrors(null, statusName.CONNECTION_REFUSED);
-    }
-  } catch (error) {
-    msg = `Error in getAllComponentWithDetailsService() function. Caused by ${error}`;
-    console.log(msg);
-    componentList = await checkErrors(error, statusName.CONNECTION_ERROR);
-  }
-  return componentList;
-};
-
-/**
- * @description get all paginated components with bipolar-transistor from the database
- * @param {any} req any type
- * @param {any} res any type
- * @returns a json object with the transaction performed
- * @example
- */
-const getAllComponentWithBipolarTransistorService = async (req, res) => {
-  try {
-    componentList = null;
-    queryStrParams = null;
-    //Pagination
-    pageSizeNro = 5;
-    pageNro = 0;
-    orderBy = 'id';
-    orderAt = 'ASC';
-    msg = null;
-
-    //-- start with pagination  ---
-    queryStrParams = req.query;
-
-    if (queryStrParams != (null && undefined)) {
-      //Pagination
-      pageSizeNro = queryStrParams.limit
-        ? parseInt(queryStrParams.limit)
-        : pageSizeNro;
-      pageNro = queryStrParams.page ? parseInt(queryStrParams.page) : pageNro;
-      orderBy = queryStrParams.orderBy ? queryStrParams.orderBy : orderBy;
-      orderAt = queryStrParams.orderAt ? queryStrParams.orderAt : orderAt;
-    }
-
-    orderBy = await checkOrderBy(orderBy);
-
-    if (orderBy == (null || undefined)) {
-      return ORDER_BY_NAME_VALUE_ERROR;
-    }
-
-    orderAt = await checkOrderAt(orderAt);
-
-    if (orderAt == (undefined || null)) {
-      return ORDER_AT_NAME_VALUE_ERROR;
-    }
-
-    order = [[orderBy, orderAt]];
-
-    //-- end with pagination  ---
-
-    if (Component != (null && undefined)) {
-      await Component.findAll({
-        attributes: {},
-        where: {
-          categoria: {
-            [Op.like]: `%Transistores BJT%`, //containing what is entered, less strictmatch
-          },
-        },
-        include: [{ model: BipolarTransistor, required: true }],
-        limit: pageSizeNro,
-        offset: pageNro,
-        order: order,
-        raw: true,
-        nest: true,
-      })
-        .then(async (componentItems) => {
-          componentList = componentItems;
-        })
-        .catch(async (error) => {
-          msg = `Error in getAllComponentWithBipolarTransistorService() function when trying to get all paginated components. Caused by ${error}`;
-          console.log(msg);
-
-          componentList = await checkErrors(error, error.name);
-        });
-    } else {
-      componentList = await checkErrors(null, statusName.CONNECTION_REFUSED);
-    }
-  } catch (error) {
-    msg = `Error in getAllComponentWithBipolarTransistorService() function. Caused by ${error}`;
-    console.log(msg);
-    componentList = await checkErrors(error, statusName.CONNECTION_ERROR);
-  }
-  return componentList;
-};
 
 /**
  * @description get all paginated components with all models from the database
@@ -1625,11 +1203,6 @@ const getAllComponentLikePriceMinMaxService = async (req, res) => {
 };
 
 module.exports = {
-  deleteComponentService,
-  getAllComponentService,
-  getAllWithAttributesComponentService,
-  getAllComponentWithDetailsService,
-  getAllComponentWithBipolarTransistorService,
   getAllComponentWithAllModelsService,
   getComponentByIdService,
   getAllComponentLikeCodeService,
